refactor: move singleton-combo helper out of workflow into utils
- Create api/utils.py as the home for shared business-logic helpers that
  span multiple api modules but are unrelated to the workflow state machine.
- Move _SHARED_GLAZE_FIELDS (renamed to SHARED_GLAZE_FIELDS) and
  sync_glaze_type_singleton_combination from api/workflow.py into
  api/utils.py. workflow.py should only contain state-machine helpers
  derived from workflow.yml.
- Update imports in api/admin.py and api/manual_tile_imports.py to pull
  from api/utils instead of api/workflow.
- Document the module-boundary rule in docs/agents/glaze-domain.md so
  future agents know not to put non-state-machine logic in workflow.py.
